+++ backend/src/engine.rs
@@ -77,6 +77,8 @@ pub async fn start(config: BackendConfig, state: StateStore) -> Result<EngineRun
     let mut player_config = PlayerConfig {
         bitrate: config.bitrate,
         position_update_interval: Some(Duration::from_secs(1)),
+        normalisation: config.normalisation,
+        normalisation_pregain_db: config.normalisation_pregain_db,
         ..PlayerConfig::default()
     };
     player_config.gapless = true;